+++ CHANGELOG.md
@@ -4,7 +4,10 @@ CHANGELOG
 ## HEAD (Unreleased)
 * using font-awesome
 * removed some unused dependecies
-* Moved Marey's trains and Hexbin examples to example sub project
+* Moved Marey's trains and Hexbin examples to [example sub project](https://github.com/Hypercubed/Project-Chi-Test)
+* Updated angular-ui-bootstrap
+* Now using npm versions of all angular modules
+* Now set `debugEnabled` and `debugInfoEnabled` based on System.production
 
 --------------------
 
@@ -16,4 +19,3 @@ _(none)_
 
 ## 1.0.0-alpha (2016-04-28)
 _(none)_
-